big merge

+++ packages/ai/src/main.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import { askAI } from './aiInterface';
 import cors from 'cors';
 import { generateAIPrompt } from './promptGenerator';
-import { saveDoc, getDoc, checkIfIDInUse, checkIfDocIsExisting, getAllDocs, deleteEverything, deleteDoc, deleteID } from './storageController';
+import { saveDoc, getDoc, checkIfIDInUse, checkIfDocIsExisting, getAllDocs, deleteEverything, deleteDoc, deleteID, getAiSource, setAiSource, getAllDocsForSpecificId, doc } from './storageController';
 
 // Erstellen einer neuen Express-Anwendung
 const app = express();
@@ -16,6 +16,38 @@ let applyedTheme = {};
 
 // Die ID mit der der User gerade "angemeldet" ist
 let currentID: string;
+
+//Ist nur true während die AI eine Anfrage bearbeitet
+let isAiLoading = false;
+
+//Das Theme welche zuletzt gelöscht wurde
+let lastDeletedThemeData : DeletedThemeData | null = null;
+
+interface DeletedThemeData {
+  id: string;
+  docName: string;
+  themeandquestionnaire: {
+    json: any;  
+    questionnaire: any;  
+  };
+}
+
+//Die ID welche zuletzt gelöscht wurde
+let lastDeletedIdData : DeletedIDData | null = null;
+
+interface Theme {
+  doc: {
+    json: object;
+    questionnaire: object;
+  };
+}
+
+interface DeletedIDData {
+  id: string;
+  themes: {
+    [key: string]: Theme;
+  };
+}
 
 // CORS und JSON Middleware verwenden
 app.use(cors({ origin: 'http://localhost:4200' }));
@@ -49,7 +81,8 @@ function hasEmptyStrings(value: unknown): boolean {
 async function retryAskAI(AIprompt: string, preferences: any, temperature: number, retries = 5) {
   for (let attempt = 1; attempt <= retries; attempt++) {
     try {
-      const response = await askAI(AIprompt, temperature);
+      const aiSourceDoc = await getAiSource();
+      const response = await askAI(AIprompt, aiSourceDoc.aiSource, temperature);
       console.log(`AI Response Attempt ${attempt}: ${response}`);
 
       let jsonMatch = null;
@@ -97,12 +130,15 @@ app.post('/api/generate-with-ai', async (req, res) => {
   const userPreferences = message;
   let AIprompt = generateAIPrompt(userPreferences);
   try {
+    isAiLoading = true;
     latestGeneratedTheme = await retryAskAI(AIprompt, userPreferences, temperature);
     applyedTheme = JSON.parse(JSON.stringify(latestGeneratedTheme));
-    res.json({ success: true, theme: latestGeneratedTheme });
+    res.json({ successFullConnectionToAi: true, theme: latestGeneratedTheme });
   } catch (error) {
     console.error('AI Request failed:', error);
-    res.status(500).json({ success: false, message: 'Failed to generate theme with AI. Please try again later.' });
+    res.json({ successFullConnectionToAi: false, message: 'Connection to AI failed' });
+  } finally {
+    isAiLoading = false;
   }
 });
 
@@ -135,16 +171,16 @@ app.post('/api/save-questionnaire', async (req, res) => {
   const data = req.body;
 
   if (!data.saveUnder) {
-    res.json({ success: false, message: "Der Name darf nicht leer sein!" })
+    res.json({ success: false, message: "Please choose a name under which to save the theme!" })
   } else if (!currentID) {
-    res.json({ success: false, message: "Die ID darf nicht leer sein!" })
+    res.json({ success: false, message: "The ID cant be empty!" })
   } else if (Object.keys(latestGeneratedTheme).length === 0) {
-    res.json({ success: false, message: "Sie müssen zuerst ihren Fragebogen abschicken! Es wird immer die zuletzt abgeschickte Fragebogen gespeichert." })
+    res.json({ success: false, message: "Please submit the questionnaire first!" })
   } else if (await checkIfDocIsExisting(currentID, data.saveUnder)) {
-    res.json({ success: false, message: "Name für diese ID bereits vergeben" })
+    res.json({ success: false, message: "Name for the theme with this ID already in use!" })
   } else {
     await saveDoc(currentID, data.saveUnder, latestGeneratedTheme, data.message)
-    res.json({ success: true, message: "Theme erfolgreich gespeichert!" })
+    res.json({ success: true, message: "Theme saved successfully!" })
   }
 });
 
@@ -190,17 +226,68 @@ app.post('/deleteDatabaseEntries', async (req, res) => {
 
 app.post('/deleteDoc', async (req, res) => {
   const data = req.body
-  const doc = await deleteDoc(data.category, data.docName)
-  res.json({ success: true })
+  const themeandquestionnaire = await getDoc(data.category, data.docName);
+  if (themeandquestionnaire){
+    lastDeletedThemeData = { id : data.category, docName : data.docName, themeandquestionnaire : { json: themeandquestionnaire.json, questionnaire: themeandquestionnaire.questionnaire}}
+    await deleteDoc(data.category, data.docName)
+    res.json({ success: true })
+  } else {
+    res.status(400).json({ success: false, ok: false, message: 'No document to delete' });
+  }
+});
+
+app.put('/undoDeleteDoc', async (req, res) => {
+  if(lastDeletedThemeData) {
+    saveDoc(lastDeletedThemeData.id.replace('O4S-ai-', ''), lastDeletedThemeData.docName, lastDeletedThemeData.themeandquestionnaire.json, lastDeletedThemeData.themeandquestionnaire.questionnaire)
+    res.status(200).json({ success: true, ok: true });
+    lastDeletedThemeData = null;
+  } else {
+    res.status(400).json({ success: false, ok: false, message: 'No document to undo delete' });
+  }
 });
 
 app.post('/deleteID', async (req, res) => {
   const data = req.body
-  const doc = await deleteID(data.category)
+  const themes = await getAllDocsForSpecificId(data.category)
+  if (data.category){
+    lastDeletedIdData = { id: data.category, themes : themes }
+    await deleteID(data.category)
+    res.status(200).json({ success: true, ok: true });
+  } else {
+    res.status(400).json({ success: false, ok: false, message: 'No ID to delete' });
+  }
+});
+
+app.put('/undoDeleteID', async (req, res) => {
+  if(lastDeletedIdData) {
+    for (const key in lastDeletedIdData.themes) {
+      const keyAsString = key;
+      saveDoc(lastDeletedIdData.id.replace('O4S-ai-', ''), keyAsString, lastDeletedIdData.themes[key].doc.json, lastDeletedIdData.themes[key].doc.questionnaire);
+    }
+    res.status(200).json({ success: true, ok: true });
+    lastDeletedIdData = null
+  } else {
+    res.status(400).json({ success: false, ok: false, message: 'No document to undo delete' });
+  }
+});
+
+app.post('/setAiSource', async (req, res) => {
+  const data = req.body
+  await setAiSource(data.aiSource)
   res.json({ success: true })
 });
 
+app.get('/getAiSource', async (req, res) => {
+  const aiSource = await getAiSource();
+  res.json(aiSource)
+});
+
+app.get('/getIsAiLoading', async (req, res) => {
+  res.json({isAiLoading : isAiLoading})
+});
+
 app.listen(PORT, () => {
+  setAiSource("local");
   console.log(`AI Backend Server running on port ${PORT}`);
 });
